+++ net/crates/net/src/adapter/net/cortex/tasks/adapter.rs
@@ -58,7 +58,7 @@ struct TasksSnapshotPayload {
 pub struct TasksAdapter {
     inner: CortexAdapter<TasksState>,
     /// Producer identity stamped on every `EventMeta`.
-    origin_hash: u32,
+    origin_hash: u64,
     /// Monotonic per-origin counter for `EventMeta::seq_or_ts`.
     /// Shared with the inner `WatermarkingFold` wrapper around
     /// [`TasksFold`]: the fold task advances this counter via
@@ -84,7 +84,7 @@ impl TasksAdapter {
     /// `app_seq` past any pre-existing same-origin `seq_or_ts`,
     /// so the first `ingest_typed` after `open` cannot collide
     /// with an already-stored event.
-    pub async fn open(redex: &Redex, origin_hash: u32) -> Result<Self, CortexAdapterError> {
+    pub async fn open(redex: &Redex, origin_hash: u64) -> Result<Self, CortexAdapterError> {
         Self::open_with_config(redex, origin_hash, RedexFileConfig::default()).await
     }
 
@@ -92,7 +92,7 @@ impl TasksAdapter {
     /// (useful for `persistent: true` or custom retention).
     pub async fn open_with_config(
         redex: &Redex,
-        origin_hash: u32,
+        origin_hash: u64,
         redex_config: RedexFileConfig,
     ) -> Result<Self, CortexAdapterError> {
         let name = ChannelName::new(TASKS_CHANNEL).map_err(|e| {
@@ -286,7 +286,7 @@ impl TasksAdapter {
     /// See [`Self::open`] for why this is `async`.
     pub async fn open_from_snapshot(
         redex: &Redex,
-        origin_hash: u32,
+        origin_hash: u64,
         state_bytes: &[u8],
         last_seq: Option<u64>,
     ) -> Result<Self, CortexAdapterError> {
@@ -304,7 +304,7 @@ impl TasksAdapter {
     /// `RedexFileConfig` (e.g. for `persistent: true`).
     pub async fn open_from_snapshot_with_config(
         redex: &Redex,
-        origin_hash: u32,
+        origin_hash: u64,
         redex_config: RedexFileConfig,
         state_bytes: &[u8],
         last_seq: Option<u64>,
